requirements.txt
discord.py
wolframalpha
requests